Work in Progress
- TODO, make unit test that is dictionary and ncache handles, no backplane. then put on an absolute expiry of 5 seconds, and see
if both ncache and dictionary are cleared? Since no backplane, i don't think it would be expired. Do we want it to be (how does redis do it with the Key thing?)?
- look at the comment on here http://cachemanager.michaco.net/documentation/CacheManagerCacheSynchronization and it mentions that in memory cache should have
much shorter duration than redis. So make unit test with in memory and 5 second expiry, and ncache with 20 second. Then see how it behaves

- Interesting, this feature https://github.com/MichaCo/CacheManager/issues/111 was added and lets it behave as just a backplane using the pub/sub mechanism, but really only
for remove (invalidate an item from other in memory caches). Not sure if we could do this??

- also implement the Retry logic which has the try/catching for ncache exceptions
- Lastly, look at the redis Update override, and see if we need an update override for ncache